chore: move platform-independent impl to header
