+++ libtransmission/file.cc
@@ -13,6 +13,7 @@
 #include "libtransmission/error.h"
 #include "libtransmission/file.h"
 #include "libtransmission/tr-assert.h"
+#include "libtransmission/tr-strbuf.h"
 
 using namespace std::literals;
 
@@ -36,14 +37,15 @@ bool tr_sys_file_write_line(tr_sys_file_t handle, std::string_view buffer, tr_er
     return ret;
 }
 
-std::vector<std::string> tr_sys_dir_get_files(std::string const& folder, std::function<bool(std::string_view)> const& test)
+std::vector<std::string> tr_sys_dir_get_files(std::string_view folder, std::function<bool(std::string_view)> const& test)
 {
     if (auto const info = tr_sys_path_get_info(folder); !info || !info->isFolder())
     {
         return {};
     }
 
-    auto const odir = tr_sys_dir_open(folder.c_str());
+    auto const szfolder = tr_pathbuf{ folder };
+    auto const odir = tr_sys_dir_open(szfolder);
     if (odir == TR_BAD_SYS_DIR)
     {
         return {};
