+++ libtransmission/file.h
@@ -615,7 +615,7 @@ char const* tr_sys_dir_read_name(tr_sys_dir_t handle, struct tr_error** error = 
 bool tr_sys_dir_close(tr_sys_dir_t handle, struct tr_error** error = nullptr);
 
 [[nodiscard]] std::vector<std::string> tr_sys_dir_get_files(
-    std::string const& folder,
+    std::string_view folder,
     std::function<bool(std::string_view)> const& test = [](auto const&) { return true; });
 
 /** @} */
